[FEAT] 게시글 작성 core 파일 분리 | Separated core files for write articles TALEARNT-402
