+++ src/pages/articles/WriteArticle/core/writeArticle.api.ts
@@ -1,21 +1,6 @@
-import { postAPI, putAPI } from "@utils/apiMethods";
+import { postAPI } from "@utils/apiMethods";
 
-import { matchingArticleType } from "@pages/articles/MatchingArticleList/core/matchingArticleList.type";
-import {
-  matchingArticleBodyType,
-  presignedURLBodyType
-} from "@pages/articles/WriteArticle/core/writeArticle.type";
-
-export const postMatchingArticle = async (body: matchingArticleBodyType) =>
-  await postAPI<number>("/v1/posts/exchanges", body, { withCredentials: true });
-
-export const putMatchingArticle = async ({
-  exchangePostNo,
-  ...body
-}: matchingArticleBodyType & Pick<matchingArticleType, "exchangePostNo">) =>
-  await putAPI(`/v1/posts/exchanges/${exchangePostNo}`, body, {
-    withCredentials: true
-  });
+import { presignedURLBodyType } from "@pages/articles/WriteArticle/core/writeArticle.type";
 
 export const postToGetPresignedURL = async (files: presignedURLBodyType[]) =>
   await postAPI<string[]>("/v1/uploads", files, { withCredentials: true });
