+++ src/pages/articles/WriteArticle/core/writeArticle.constants.ts
@@ -1,50 +1,12 @@
 import { array, mixed, number, object, string } from "yup";
 
-import { CATEGORIZED_TALENTS_LIST } from "@common/common.constants";
-
 import {
-  durationType,
-  exchangeType,
-  postType
-} from "@pages/articles/core/articles.type";
+  durationList,
+  exchangeTypeList,
+  postTypeList
+} from "@pages/articles/core/articles.constants";
 
-export const articleTypeOptions = [
-  { label: "매칭 게시물 글쓰기", value: "match" },
-  { label: "커뮤니티 게시물 글쓰기", value: "community" }
-];
-export const durationList: durationType[] = [
-  "기간 미정",
-  "1개월",
-  "2개월",
-  "3개월",
-  "3개월 이상"
-];
-export const exchangeTypeList: exchangeType[] = [
-  "온라인",
-  "오프라인",
-  "온/오프라인"
-];
-export const postTypeList: postType[] = [
-  "자유 게시판",
-  "질문 게시판",
-  "스터디 모집 게시판"
-];
-
-export const durationOptions = durationList.map(item => ({
-  label: item,
-  value: item
-}));
-export const talentsOptions = CATEGORIZED_TALENTS_LIST.map(
-  ({ categoryName, talents }) => ({
-    label: categoryName,
-    value: talents.map(({ talentCode, talentName }) => ({
-      label: talentName,
-      value: talentCode
-    }))
-  })
-);
-
-export const matchArticleSchema = object({
+export const matchingArticleSchema = object({
   giveTalents: array()
     .of(number().required())
     .min(1, "재능 키워드를 선택해 주세요")
@@ -59,6 +21,13 @@ export const matchArticleSchema = object({
     .required("제목을 입력해 주세요")
     .min(2, "제목을 2글자 이상 입력해 주세요"),
   content: string().required(),
+  pureText: string()
+    .required("내용을 입력해 주세요")
+    .test(
+      "min-length",
+      "내용을 20글자 이상 입력해 주세요",
+      value => value.length >= 21
+    ),
   imageFileList: array()
     .of(
       object({
@@ -69,14 +38,7 @@ export const matchArticleSchema = object({
         url: string().required()
       }).required()
     )
-    .required(),
-  pureText: string()
-    .required("내용을 입력해 주세요")
-    .test(
-      "min-length",
-      "내용을 20글자 이상 입력해 주세요",
-      value => value.length >= 21
-    )
+    .required()
 }).required();
 export const communityArticleSchema = object({
   postType: string().oneOf(postTypeList).required(),
@@ -84,6 +46,13 @@ export const communityArticleSchema = object({
     .required("제목을 입력해 주세요")
     .min(2, "제목을 2글자 이상 입력해 주세요"),
   content: string().required(),
+  pureText: string()
+    .required("내용을 입력해 주세요")
+    .test(
+      "min-length",
+      "내용을 20글자 이상 입력해 주세요",
+      value => value.length >= 21
+    ),
   imageFileList: array()
     .of(
       object({
@@ -94,12 +63,5 @@ export const communityArticleSchema = object({
         url: string().required()
       }).required()
     )
-    .required(),
-  pureText: string()
-    .required("내용을 입력해 주세요")
-    .test(
-      "min-length",
-      "내용을 20글자 이상 입력해 주세요",
-      value => value.length >= 21
-    )
+    .required()
 }).required();
